// src/style/foundation/typo/typo.type.ts
export type TypoKRType =
  | 'D1_KR_Sb_96'
  | 'D1_KR_Rg_96'
  | 'D1_KR_Lt_96'
  | 'D2_KR_Sb_88'
  | 'D2_KR_Rg_88'
  | 'D2_KR_Lt_88'
  | 'D3_KR_Sb_80'
  | 'D3_KR_Lt_80'
  | 'D3_KR_Rg_80'
  | 'D4_KR_Sb_72'
  | 'D4_KR_Rg_72'
  | 'D4_KR_Lt_72'
  | 'D5_KR_Sb_64'
  | 'D5_KR_Rg_64'
  | 'D5_KR_Lt_64'
  | 'D6_KR_Sb_56'
  | 'D6_KR_Rg_56'
  | 'D6_KR_Lt_56'
  | 'H1_KR_Sb_48'
  | 'H1_KR_Rg_48'
  | 'H1_KR_Lt_48'
  | 'H2_KR_Sb_40'
  | 'H2_KR_Rg_40'
  | 'H2_KR_Lt_40'
  | 'H3_KR_Sb_32'
  | 'H3_KR_Rg_32'
  | 'H3_KR_Lt_32'
  | 'T1_KR_Sb_28'
  | 'T1_KR_Rg_28'
  | 'T1_KR_Lt_28'
  | 'T2_KR_Sb_24'
  | 'T2_KR_Rg_24'
  | 'T2_KR_Lt_24'
  | 'T3_KR_Sb_20'
  | 'T3_KR_Rg_20'
  | 'T3_KR_Lt_20'
  | 'T4_KR_Sb_18'
  | 'T4_KR_Rg_18'
  | 'T4_KR_Lt_18'
  | 'B1_KR_Sb_16'
  | 'B1_KR_Rg_16'
  | 'B1_KR_Lt_16'
  | 'B2_KR_Sb_15'
  | 'B2_KR_Rg_15'
  | 'B2_KR_Lt_15'
  | 'B3_KR_Sb_14'
  | 'B3_KR_Rg_14'
  | 'B3_KR_Lt_14'
  | 'C1_KR_Sb_13'
  | 'C1_KR_Rg_13'
  | 'C1_KR_Lt_13'
  | 'C2_KR_Sb_12'
  | 'C2_KR_Rg_12'
  | 'C2_KR_Lt_12'
  | 'C3_KR_Sb_11'
  | 'C3_KR_Rg_11'
  | 'C3_KR_Lt_11';

export type TypoENType =
  | 'D1_EN_Sb_96'
  | 'D1_EN_Rg_96'
  | 'D1_EN_Lt_96'
  | 'D2_EN_Sb_88'
  | 'D2_EN_Rg_88'
  | 'D2_EN_Lt_88'
  | 'D3_EN_Sb_80'
  | 'D3_EN_Rg_80'
  | 'D3_EN_Lt_80'
  | 'D4_EN_Sb_72'
  | 'D4_EN_Rg_72'
  | 'D4_EN_Lt_72'
  | 'D5_EN_Sb_64'
  | 'D5_EN_Rg_64'
  | 'D5_EN_Lt_64'
  | 'D6_EN_Sb_56'
  | 'D6_EN_Rg_56'
  | 'D6_EN_Lt_56'
  | 'H1_EN_Sb_48'
  | 'H1_EN_Rg_48'
  | 'H1_EN_Lt_48'
  | 'H2_EN_Sb_40'
  | 'H2_EN_Rg_40'
  | 'H2_EN_Lt_40'
  | 'H3_EN_Sb_32'
  | 'H3_EN_Rg_32'
  | 'H3_EN_Lt_32'
  | 'T1_EN_Sb_28'
  | 'T1_EN_Rg_28'
  | 'T1_EN_Lt_28'
  | 'T2_EN_Sb_24'
  | 'T2_EN_Rg_24'
  | 'T2_EN_Lt_24'
  | 'T3_EN_Sb_20'
  | 'T3_EN_Rg_20'
  | 'T3_EN_Lt_20'
  | 'T4_EN_Sb_18'
  | 'T4_EN_Rg_18'
  | 'T4_EN_Lt_18'
  | 'B1_EN_Sb_16'
  | 'B1_EN_Rg_16'
  | 'B1_EN_Lt_16'
  | 'B2_EN_Sb_15'
  | 'B2_EN_Rg_15'
  | 'B2_EN_Lt_15'
  | 'B3_EN_Sb_14'
  | 'B3_EN_Rg_14'
  | 'B3_EN_Lt_14'
  | 'C1_EN_Sb_13'
  | 'C1_EN_Rg_13'
  | 'C1_EN_Lt_13'
  | 'C2_EN_Sb_12'
  | 'C2_EN_Rg_12'
  | 'C2_EN_Lt_12'
  | 'C3_EN_Sb_11'
  | 'C3_EN_Rg_11'
  | 'C3_EN_Lt_11';

export type FontWeight = 'Light' | 'Regular' | 'SemiBold';